refactor : 테스트 환경 수정
