+++ src/test/java/com/ssafy/ltw/domain/article/model/dao/ArticleDaoTest.java
@@ -1,7 +1,10 @@
 package com.ssafy.ltw.domain.article.model.dao;
 
+import com.ssafy.ltw.domain.article.model.Article;
 import com.ssafy.ltw.domain.attraction.model.AttractionInfo;
 import com.ssafy.ltw.domain.attraction.model.dao.AttractionDaoImpl;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -13,7 +16,31 @@ import static org.junit.jupiter.api.Assertions.*;
 
 class ArticleDaoTest {
     private final ArticleDao articleDao = ArticleDaoImpl.getArticleDao();
+    // TODO : 유저 구현시 수정해야함
+    @BeforeEach
+    void beforeEach(){
+    }
 
+    @AfterEach
+    void afterEach(){
+        articleDao.clear();
+    }
+    @Test
+    @DisplayName("게시물 등록한다")
+    public void 게시물_등록() throws Exception {
+        //Given
+        Article article = new Article().builder()
+                .subject("test")
+                .content("testContente")
+                .memberId(1)
+                .build();
+
+        //When
+        int res = articleDao.writeArticle(article);
+
+        //Then
+        assertThat(res).isEqualTo(1);
+    }
     @Test
     @DisplayName("전체 article 수 조회 테스트")
     public void article_count_test() throws Exception {
