+++ src/test/java/com/ssafy/ltw/domain/article/model/service/ArticleServiceTest.java
@@ -1,0 +1,19 @@
+package com.ssafy.ltw.domain.article.model.service;
+
+import com.ssafy.ltw.domain.article.model.Article;
+import com.ssafy.ltw.domain.article.model.dao.ArticleDao;
+import com.ssafy.ltw.domain.article.model.dao.ArticleDaoImpl;
+import org.junit.jupiter.api.AfterEach;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.jupiter.api.Assertions.*;
+
+class ArticleServiceTest {
+    // 레포 테스르를 하는중 수정 필요
+    private final ArticleService articleService = ArticleServiceImpl.getArticleService();
+
+
+}
